[cache] Pipeline cache to avoid `writeable` arguments
Writing to an `arg` value results in a `!torch.tensor` argument type
which violates write semantics. As scatters  are inherently mutative
operations on the destination buffer we can internally create a copy
then scatter to this copy. This should result in an in-place mutative
operation as the dispatch takes ownership of the passed buffer.