Add test script and workflow for llama export, compile, serve. (#70)
Progress on https://github.com/nod-ai/sharktank/issues/22

Sample runs on my fork:
* https://github.com/ScottTodd/sharktank/actions/runs/9670685134
* https://github.com/ScottTodd/sharktank/actions/runs/9715408887

I decided to run this on a nightly `schedule` and on
`workflow_dispatch`. It takes around 10 minutes so it _could_ run on
`pull_request` if we want too.

As these components stabilize and we spend less time hacking on
individual steps using the full toolkit (python -> manual `iree-compile`
vs. using the in-process compiler API) we can switch the test from a
bash script to a pytest file. Need to start somewhere :)