Allow sharktank tensors to be part of export function signatures
Adds flattening and unflattening of our tensor types according to torch
requirements.
This allows us to pass and return such tensors when
exporting functions.

Add function to populate names of tensors in a theta with the natural
name corresponding to its path in the theta.
When exporting a dataset each tensor must have a unique name.

Make name a property in InferenceTensor.
Renaming of sharded tensors needs to rename the shards as well.