[punet] Implement QKV quantization fusion and adapt to Brevitas mechanism.
This patch enables quantization activation fusion of either qkv (for MHA) or kv (for MCHA) attention layers.

In the model, this is selected based on the presence of either a `to_kv` or `to_qkv` theta on each attention layer. Normally, a LinearLayer handles all quantization inside of itself, but in this case, we decompose what the LinearLayer usually does by explicitly performing activation quantization using either qkv or kv shared input conditioning, then proceed with the normal computation and dequant.

The Brevitas import is a bit finnicky in this case because the current version of it concats the kv or qkv weights/scales/bias/zp in an attempt to only have one set of each. However, it is a more natural representation (and leaves less room for optimization failures) to simply populate the dataset with the data at rest how we intend to use it. So we reorganize into this structure:

```
attn0.to_qkv
  premul_input
  qinput
  to_q.weight
  to_q.bias
  to_k.weight
  to_k.bias
  to_v.weight
  to_v.bias
```

(to_kv is similar but does not include the q within it, leaving it at the top, unfused level)

We may want to switch the quantizer to a more natural (for us) representation.