[shortfin] Start building out serving demo spike. (#129)
This is the first step in a spike to add minimal features and API for
standing up a mobilenet server from an ONNX model in the ONNX model zoo,
and it is being used to build out basic features:

* Make Worker monomorphic and add a mechanism so that the Python
bindings can stash needed state and get lifecycle events that make
interop possible (without subclassing). This allows us to move the
current worker TLS lookup to the C++ side, where the rest of the
framework can use it.
* Adds basic bindings for `ProgramModule` and `Program`, allowing
loading from VMFB and setting up with all devices in a scope.
* Adds a `Message` base class that will be used for signaling between
processes. A little tricky because of the vague ownership and
binding/extension needs.
* Adds a `Queue` class with very simple support for readers and writers.
(This will be extended a lot later to support fancy features like
back-pressure shutdown sequencing, etc)
* Adds a `Future` class that can be resolved with a concrete type or
exception. `VoidFuture` is hand-coded. Others are via `TypedFuture`
template. Implement `MessageFuture`.
* Tie `Future` into asyncio so it is awaitable.
* Add script to build demo model and scaffolding for a server.

There is some torturous code in some of this because there are only so
many ways to make it idiomatic C++ and bindable to Python. There are
some shutdown and exception reporting issues still to work out (which is
typical of this state of implementation).