[shortfin] Bridge shortfin processes to FastAPI.
This is just demonstrating with an example for the moment but we can bake out the FastAPIResponder to a helper class.

The basic idea is that the shortfin world is one of devices, topologies, light-weight processes and queues, while the web server world is (mostly) single-threaded I/O loops. By letting us package a FastAPI (i.e. starlette/uvicorn) request/response as a message on a queue and disconnect it from the request/response cycle, we can perform arbitrary workflows on it and respond either one-shot or streaming as we go without worrying overly much about the different threading models, etc.