[shortfin] Build out basic array interop. (#150)
* In C++, `device_array` subclasses can now interop with xtensor
nd-array views. This is used to make the `to_s` / `__repr__` method
pretty print contents for now.
* C++ and Python APIs for mapping host memory.
* Python `storage.map()` if used for explicit mapping (with arguments to
control details) whereas the `storage.data` property maps as read-only
on getattr and as write/discard on setattr. Combined with the buffer
protocol, this makes mapping work gluelessly with other Python buffer
based objects.

The xtensor_bridge is relatively complex because it has to bridge from
the polymorphic dtype world of base_array to the static element type
world of xtensor. It does this by having an array mixin class which
lazily constructs an indirection trampoline object that is (private to
the xtensor_bridge TU) template instantiated for each supported concrete
C++ type. When the trampoline is needed, memory is mapped and the
trampoline is constructed, if possible. Because of the type-erasure and
desire to minimize per-array overhead for the case when the feature is
not used, this involves two layers of in-place construction to arrive at
the concrete xtensor adaptor. This is enough to implement basic things
like I/O. A bit more will be needed to add capabilities from there.

C++ code which knows what types it is operating on can also directly
instantiate an adaptor, but even then, it is good to have some generic
array support (like I/O, fill, tranpose, etc).

This was organized so the feature could be ifdef'd out if not desired.
In a pure C++ static library, it should just be dropped by the linker if
not used, but it does add compile time and complexity. Current cost is
about 175KiB for the xtensor bridge (out of a total library size of
1.5MiB). This will increase somewhat as more features are used.