[shortfin] Make TimelineResource hold scopes alive. (#157)
Prior to this, TimelineResource held a raw C++ reference to the scope
but did not hold a shared_ptr ref to it. The storage class, which holds
buffers alive, was keeping the raw IREE buffer and device alive but not
the shortfin object hierarchy. On the happy path, this does not cause
issues, but during abnormal termination and other race conditions, it
can cause lifetime problems.

Here we:

* Make TimelineResource hold a std::shared_ptr<Scope>, which
transitively keeps everything alive that can be accessed.
* Clean up incorrect ordering of fields in key classes, which could
cause dependent objects to be destroyed out of order.
* Remove cases in iree::object_ptr which could have resulted in
incorrect accounting in certain scenarios.
* Adds compile-time flags to perform verbose shortfin lifetime logging
and application-side IREE reference checks.
* Change System::Shutdown() to not clear object references that still
may be live (leave that for the destructor).
* Correct an issue where drivers could be retained forever, which was
then masking another lifetime issue during abnormal termination.
* Find a low rate shutdown race triggering use-after-free in the
BlockingExecutor and fix it (found by lifetime logging).
* Ensure that all tests are ASAN and LSAN clean, even with a new
abnormal termination case added (see updates to include buffer copying,
which is currently triggering a buffer permission exception, which set
this whole triage session in motion).