[libshortfin] Add missing scheduler dependency.
As written, command buffers could be scheduled out of order (with signal < deps). We need a more enlightened scheduler soon, but for the moment, this makes it correct.

Also sets the IREE runtime IREE_TASK_SCOPE_FLAG_ABORT_ON_FAILURE in debug builds, which will cause an eager assert-like crash. Since in our usage, this is always a critical error, it makes sense to treat it like an assert (and I would like to add finer grained control in the future). In release builds, it will propagate as a soft error, but it is basically fatal as it will lock the accounting structures into a persistent error state.