[libshortfin] Initial implementation of LLM inference server. (#181)
This is very much a first draft that needs a fair bit of work to turn
into a final form. As the first real user of `libshortfin`'s core APIs,
a number of rough spots were worked out even to get it to this first
stage.

Just a few things that need attention:

* Sampling is just a greedy decode loop right now, but this can be
easily extended because it is done per item and is just linear Python
(async).
* Scaffolding for multi device in a single process is in place but we're
still working on the models to mate up with it.
* A lot of configuration is hardcoded.
* The batcher is a toy. The aim is to implement something like sglang's
cache aware batcher and radix attention in its place. Given that, no
effort was taken to make the batcher more than write-once code.
* There are many opportunities to use different shortfin fibers, workers
and executors to better balance the host-side activity. This currently
just runs async on a single worker for the moment.