[libshortfin] Implement multi-device CPU system config. (#212)
On NUMA systems, a NUMA node maps to a device queue, which manifests as
a Shortfin Device. On non-NUMA systems, options can be used to control
the concurrency.

Also drops the `Device::node_locked` field and replaces it with
`capabilities`, which is needed in order to create better data transfer
APIs.