[libshortfin] Add a workflow with disabled GIL
This adds a seperate, and therefore temporary, workflow to run with a
free-threaded Python which should be embedded into a matrix strategy.

To fully enable testing without the GIL, it is required to declare that
`_shortfin_default.lib` can either be safley run without it or to run with
`PYTHON_GIL=0`.