[shortfin] Fix issues with free threaded Python builds. (#226)
* Fixes a method signature issue that manifested on 3.13 generally (not
FT related).
* Bumps nanobind to a recent HEAD commit to pick up free threaded
support.
* Adds nanobind option to enable free threading if building for a
CPython with it enabled.
* Adds a README stanza advising on how to acquire a free threaded
CPython.
* Makes fastapi tests skip if deps not met (not yet available / hard to
install).
* Adds an LSAN exclusion for something unrelated that has snuck in.
* Sets FT CMAKE_BUILD_TYPE=Debug in the CI.

Note that on the large NUMA system I was testing on, the CPU test which
creates an executor on each NUMA node for all processes exceeded the
default file handle ulimit, requiring it to be increased. I assume that
the FT CPython internally uses more synchronization handles and it just
happened to go over budget. This resulted in fixing
https://github.com/iree-org/iree/pull/18609, which was causing an assert
to be hit in this specific RESOURCE_EXHAUSTED scenario.