[shortfin] Make system construction selectable by type name. (#229)
Prior usage required callers to switch to know which SystemBuilder class
to create. That option of construction is still available and has more
explicitness.

This patch adds C++ APIs for construction of a SystemBuilder by
type_name and a shorthand on the System class to have a one line create.
Also reworks the config undef checking to be consistent and make it
possible to warn instead of throw an exception.

In Python, this is exposed by the System class getting a constructor, so
now a one-liner like `with sf.System("hostcpu", option1="value", ...):`
is possible. This should compose well with flags and application code.