Introduce CausalLMModel intefrace and add IREE numerics test for Llama 3.1 8B FP16 TP8
We do not have a clearly defined interface for LMs decode and prefill
have different signature when exporting to IREE.
Here is added a new ABC CausalLMModel that makes a distinction between
the to variants.
The BaseCausalLMModel provides a default implementation for the new
prefill_from_seq_lens and decode_from_seq_lens methods.

The export script export_paged_llm_v1 does too much in its exported
functions, first it computes the attention mask then it shards its
arguments results.
This change moves this into a separate class that conforms to the
CausalLMModel intefrace.

Introduce a new CausalLMIreeModel that conforms to CausalLMModel, but is
backed by an IREE module.
It is not performant and only meant for testing as it marshals tensors
and uses the IREE Python bindings.
This can then be used for example in the paged_llm_v1.TorchGenerator
or other places where a LM is expected.

Refactor the sharded Llama tests. Increase code reuse and use the
TorchGenerator in the toy-sized tests. Use the shard_llm_dataset and
export_paged_llm_v1 scripts in the test flow to increase their test
coverage.

Introduce a Llama 3.1 8B FP16 TP8 test that appears to not have good
numerical accuracy. It is compared to an fp64 unsharded torch variant
to ensure that the reference is of high accuracy.