Revise and split requirements files
Moves dependencies defined in the root `requirements.txt` to
`sharktank/` and splits out test only deps to `requirements-tests.txt`
file. Dependencies only used for development / in the CI are moved to
`requirements-dev.txt` and the root requirements file is now used to
pull in all the deps. Furthermore, some no longer used dependencies
are removed.