Apply new versioning scheme
Progress on #130 and #294.

With this, every package has its own `version_info.json` which defaults
to a dev build (`X.Y.Z.dev`). Based on this version information the
version identifier for a release candate can be computed
(`X.Y.ZrcYYYYMMDD`) which is written to corresponding
`version_info_rc.json` files. This PR further adapt the build packages
workflow to make use of the script and apply the new versioning scheme.
Even though sharktank packages are not yet build, the workflow already
generates the rc version numbers for sharttank and shortfin.