Add T5 encoder exporting to MLIR and numerics verification with IREE
This change adds the export functionality and a test for the T5 v1.1
encoder. Small and XXL variants.

Added parsing of GGUF arrays when loading properties. This is used to
deduce the vocabulary size instead of having to provide this manually.
This can potentially be used also to load the tokenizer from GGUF or
IRPA.