Build nightly shark-ai -package (#535)
Refactors the scripts so that a version suffix can be passed as an
argument. This is used to generate the suffix only once in the workflow
instead of computing it in multiple scripts and partly scattered across
different steps (`write_requirements.py` runs in a different step).
Further adds dependencies to IREE packages also for nightly shark-ai
builds.

Requires #519.