[shortfin] Add C++ tokenizer wrapper library. (#610)
* This is gated by SHORTFIN_ENABLE_TOKENIZERS (presently off).
* I'd like to either take over the wrapper or get
https://github.com/mlc-ai/tokenizers-cpp/issues/50 before putting much
weight on this.
* There is no great C++ option for this component, so we go to the
trouble of integrating a Rust component. We will need to do a bit of
prep on our CI systems to enable this by default.
* Python API will be added in a subsequent commit. This should be more
efficient than the tokenizers Python API since we will allow direct
access to the tokens vs doing a lot of conversions.
* Size analysis: Prior to this patch, libshortfin was 1.8MB, which gave
us an entire GPU and CPU runtime stack. After this patch (stripped) it
is 8.4MB. Given how important the use case is, I'm willing to tolerate
this for the moment. It seems like there is room for something better
here, which is why I did not expose the underlying vendor'd API directly
(edit: by switching to a nightly rust and activating a bunch of options
from https://github.com/johnthagen/min-sized-rust, I was able to produce
a binary that is 4.2MB, which is more reasonable).