[sharktank] Explicitly install nightlies
As `iree-turbine` is specified in `sharktank/requirements.txt`, the
dependency is already fulfilled and nightlies don't get installed. One
option would be to enforce to `--upgrade` but instead the packages get
not installed before stable releases get pulled in via sharktank's
requirements file. For workflow running on non-GH hosted runners,
`--upgrade` is passed additionaly to avoid using an eventually cached
version on the self-hosted runners.